A complete transition to JDBC
[+] - Removed all the unnecessary JPA implementations.
[+] - Simplified the backend structure.
