--- backend/src/main/java/com/synnigeograafia/backend/repository/PersonRepository.java
@@ -1,11 +1,0 @@
-package com.synnigeograafia.backend.repository;
-
-import com.synnigeograafia.backend.domain.Person;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-import java.util.UUID;
-
-@Repository
-public interface PersonRepository extends JpaRepository<Person, UUID> {
-}
